Create Maximum 69 Number
Algorithm : 🧑‍💻
Step 1 : Start the function . 
Step 2: Declare the variable max to store given num value as the maximum value. 
Step 3 : To convert the single number into string to check the number is 9 or 6 . 
Step 4 : If the starting index to check the index value one by one using the for loop . 
Step 5 : Change the value 9 as 6 and 6 as 9 and then check the maximum value by converting the string to integer inside the for loop. 
Step 6 : Finally return the maximum value in output by change at most one time as 6 and 9 
Step 7 : Stop the function . 

Conclusion : 📃
  Only change the at most one time to 9 as 6 and 6 as 9 to find the maximum value . 

Real time usage : 🤏
 *Example: "Change one digit in your reward code to maximize your discount."